Fix issue with floating points not added to info update.

+++ src/WifiHandler.cpp
@@ -66,9 +66,9 @@ void WifiHandler::handleEdit()
 void WifiHandler::handleSubmit()
 {
     uint16_t pid = std::stoi(m_server.arg("pid").c_str());
-    double minValue = std::stoi(m_server.arg("minValue").c_str());
-    double maxValue = std::stoi(m_server.arg("maxValue").c_str());
-    double increment = std::stoi(m_server.arg("increment").c_str());
+    double minValue = std::stod(m_server.arg("minValue").c_str());
+    double maxValue = std::stod(m_server.arg("maxValue").c_str());
+    double increment = std::stod(m_server.arg("increment").c_str());
     uint64_t pace = std::stoi(m_server.arg("pace").c_str());
 
     auto entry = m_obdHandler->getByPid(pid);
